requirements.txt
flask
diskcache
gunicorn
peewee
django-environ